misra; fix essential type for char literals

+++ addons/test/misra/misra-test.c
@@ -643,10 +643,10 @@ static void misra_10_1_ternary(void)
 }
 
 static void misra_10_2(void) {
-    unsigned int u8a = 0;
-    signed char cha = 0;
-    signed int s8a = 0;
-    signed short s16a = 0;
+    uint8_t u8a = 0;
+    char cha = 0;
+    int8_t s8a = 0;
+    int16_t s16a = 0;
     float f32a = 0.0;
     char res;
 
@@ -654,9 +654,9 @@ static void misra_10_2(void) {
     res = s8a + '0';
     res = cha - '0';
     res = '0' - s8a;
-    res = cha + ':';
+    res = cha + ':'; // 10.2
 
-    res = s16a - 'a'; // 10.2
+    res = s16a - 'a'; // 10.2 10.3
     res = '0' + f32a; // 10.2 10.4
 }
